3008: Hackily fixed dependency cycle between (server-boilerplate <-> entity-server) - Need to work out a proper solution
