+++ packages/entity-server/package.json
@@ -19,7 +19,7 @@
     "lint:all": "yarn run lint \"src/**/*.{ts,jsx}\"",
     "lint:fix": "yarn lint --fix",
     "prestart": "npm run -s build",
-    "start": "node dist",
+    "start": "node dist/main.js",
     "start-dev": "../../scripts/bash/backendStartDev.sh 9996 -ts",
     "start-verbose": "LOG_LEVEL=debug npm run start-dev",
     "test": "jest",
